Moving twitch to its own command module - still have debug code in weather at this point
