+++ .gitignore
@@ -1,0 +1,4 @@
+node_modules/
+package.json
+package-lock.json
+slack-irc/lib/apikeys.js
